添加注释

+++ Hardware/Motor.h
@@ -1,4 +1,4 @@
 #ifndef __MOTOR_H
 void Motor_Init();
-void Motor_SetSpeed(int16_t speed);
+void Motor_SetSpeed(int8_t speed);
 #endif // !__MOTOR_H
